requirements-dev.txt
# Development dependencies for DShield MCP
# Install with: pip install -r requirements-dev.txt

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# Code formatting and linting
black>=24.3.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.0.0

# Type checking
types-python-dateutil>=2.8.0

# Documentation
sphinx>=6.0.0
sphinx-rtd-theme>=1.2.0
pdoc>=14.0.0
pydoc-markdown>=4.8.0

# Development tools
pre-commit>=3.0.0
tox>=4.0.0

# Jupyter notebooks (for development/testing)
jupyter>=1.0.0
ipykernel>=6.0.0

# Debugging
ipdb>=0.13.0

# TUI Development and Testing
textual-dev>=0.1.0  # Development tools for textual
textual-pilot>=0.1.0  # Integration testing for textual apps
rich-cli>=1.0.0     # CLI tools for rich
pytest-textual-snapshot>=0.1.0  # Snapshot testing for textual apps

# Security and Testing Tools
bandit>=1.7.0        # Security linting
safety>=2.0.0        # Security vulnerability scanning

requests>=2.32.4 # not directly required, pinned by Snyk to avoid a vulnerability
setuptools>=78.1.1 # not directly required, pinned by Snyk to avoid a vulnerability
urllib3>=2.5.0 # not directly required, pinned by Snyk to avoid a vulnerability
zipp>=3.19.1 # not directly required, pinned by Snyk to avoid a vulnerability
